# services/tanren-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tanren",
    "description": "Tanren worker-manager HTTP API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/vm": {
      "get": {
        "tags": [
          "vm"
        ],
        "summary": "List Vms",
        "description": "List active VM assignments.",
        "operationId": "list_vms_api_v1_vm_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VMSummary"
                  },
                  "title": "Response List Vms Api V1 Vm Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vm/provision": {
      "post": {
        "tags": [
          "vm"
        ],
        "summary": "Provision Vm",
        "description": "Provision a new VM.\n\nReturns:\n    VMHandle-shaped dict with vm_id, host, provider, created_at.",
        "operationId": "provision_vm_api_v1_vm_provision_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Provision Vm Api V1 Vm Provision Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vm/{vm_id}": {
      "delete": {
        "tags": [
          "vm"
        ],
        "summary": "Release Vm",
        "description": "Release a VM assignment.",
        "operationId": "release_vm_api_v1_vm__vm_id__delete",
        "parameters": [
          {
            "name": "vm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "VM identifier",
              "title": "Vm Id"
            },
            "description": "VM identifier"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMReleaseConfirmed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vm/dry-run": {
      "post": {
        "tags": [
          "vm"
        ],
        "summary": "Dry Run Provision",
        "description": "Dry-run provision \u2014 show what would happen without creating resources.",
        "operationId": "dry_run_provision_api_v1_vm_dry_run_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMDryRunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProvisionRequest": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project",
            "description": "Project name"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "description": "Git branch"
          },
          "environment_profile": {
            "type": "string",
            "title": "Environment Profile",
            "description": "Environment profile",
            "default": "default"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project",
          "branch"
        ],
        "title": "ProvisionRequest",
        "description": "Request to provision a VM \u2014 subset of Dispatch fields."
      },
...
      "VMDryRunResult": {
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/VMProvider",
            "description": "VM provider that would be used"
          },
          "server_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Type",
            "description": "Server type that would be used"
          },
          "estimated_cost_hourly": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Hourly",
            "description": "Estimated hourly cost"
          },
          "would_provision": {
            "type": "boolean",
            "title": "Would Provision",
            "description": "Whether provisioning would proceed"
          },
          "requirements": {
            "$ref": "#/components/schemas/VMRequirements",
            "description": "Resolved VM requirements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "would_provision",
          "requirements"
        ],
        "title": "VMDryRunResult",
        "description": "Result of a VM dry-run provisioning check."
      },
      "VMProvider": {
        "type": "string",
        "enum": [
          "manual",
          "hetzner"
        ],
        "title": "VMProvider",
        "description": "Supported VM providers."
      },
...
      "VMReleaseConfirmed": {
        "properties": {
          "vm_id": {
            "type": "string",
            "title": "Vm Id",
            "description": "Released VM identifier"
          },
          "status": {
            "$ref": "#/components/schemas/VMStatus",
            "description": "Release status",
            "default": "released"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "vm_id"
        ],
        "title": "VMReleaseConfirmed",
        "description": "Confirmation that a VM was released."
      },
...
      "VMRequirements": {
        "properties": {
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "cpu": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Cpu",
            "default": 2
          },
          "memory_gb": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Memory Gb",
            "default": 4
          },
          "gpu": {
            "type": "boolean",
            "title": "Gpu",
            "default": false
          },
          "server_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Type"
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Labels"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "profile"
        ],
        "title": "VMRequirements",
        "description": "Resource requirements for VM provisioning."
      },
      "VMStatus": {
        "type": "string",
        "enum": [
          "active",
          "provisioning",
          "releasing",
          "released"
        ],
        "title": "VMStatus",
        "description": "Lifecycle status of a VM."
      },
      "VMSummary": {
        "properties": {
          "vm_id": {
            "type": "string",
            "title": "Vm Id",
            "description": "VM identifier"
          },
          "host": {
            "type": "string",
            "title": "Host",
            "description": "VM hostname or IP"
          },
          "provider": {
            "$ref": "#/components/schemas/VMProvider",
            "description": "VM provider"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id",
            "description": "Associated workflow ID"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project",
            "description": "Associated project name"
          },
          "status": {
            "$ref": "#/components/schemas/VMStatus",
            "description": "Current VM status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO 8601 creation timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "vm_id",
          "host",
          "provider",
          "status",
          "created_at"
        ],
        "title": "VMSummary",
        "description": "Summary of a VM assignment."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}